Seperate router configuration from application main entry file
